Specify data types

+++ src/consts.rs
@@ -1,8 +1,8 @@
-pub const MIDI_EVENT_NOTEON: u32 = 0;
-pub const MIDI_EVENT_NOTEOFF: u32 = 1;
-pub const MIDI_EVENT_ALLNOTESOFF: u32 = 2;
-pub const MIDI_EVENT_ALLNOTESKILLED: u32 = 3;
-pub const MIDI_EVENT_RESETCONTROL: u32 = 4;
-pub const MIDI_EVENT_CONTROL: u32 = 5;
-pub const MIDI_EVENT_PROGRAMCHANGE: u32 = 6;
-pub const MIDI_EVENT_PITCH: u32 = 7;
+pub const MIDI_EVENT_NOTEON: u16 = 0;
+pub const MIDI_EVENT_NOTEOFF: u16 = 1;
+pub const MIDI_EVENT_ALLNOTESOFF: u16 = 2;
+pub const MIDI_EVENT_ALLNOTESKILLED: u16 = 3;
+pub const MIDI_EVENT_RESETCONTROL: u16 = 4;
+pub const MIDI_EVENT_CONTROL: u16 = 5;
+pub const MIDI_EVENT_PROGRAMCHANGE: u16 = 6;
+pub const MIDI_EVENT_PITCH: u16 = 7;
